Simplify running window and recenter inside of extra step
Extract step used to take window and recenter steps as inputs to run
method and use their info to repeat steps in extraction.
Steps are still repeated, but now the attributes are passed in the
pipeline, and the core "extraction" loop is not repeated for each
scenario (value of arguments is enough to control extraction behaviour).